refactor: Add explicit file paths/extensions to all source files in flip-toolkit to appease TS
