+++ packages/flip-toolkit/package.json
@@ -8,6 +8,10 @@
   "umd:main": "lib/index.umd.js",
   "module": "lib/index.es.js",
   "types": "./lib/index.d.ts",
+  "exports": {
+    "types": "./lib/index.d.mts",
+    "import": "./lib/index.modern.mjs"
+  },
   "repository": {
     "type": "git",
     "url": "https://github.com/aholachek/react-flip-toolkit"
@@ -24,7 +28,7 @@
     "npm": ">=5"
   },
   "scripts": {
-    "build": "microbundle --name=FlipToolkit --define process.env.NODE_ENV=production --globals rematrix=Rematrix",
+    "build": "microbundle --name=FlipToolkit --define process.env.NODE_ENV=production --globals rematrix=Rematrix && cp lib/index.d.ts lib/index.d.mts",
     "build:debug": "microbundle --name=FlipToolkit --no-compress",
     "lint": "echo \"noop\"",
     "test": "echo \"noop\"",
