+++ packages/flip-toolkit/src/flip/animateUnflippedElements/types.ts
@@ -1,4 +1,4 @@
-import { BaseFlipArgs } from '../types'
+import { BaseFlipArgs } from '../types.js'
 
 export interface AnimateUnflippedElementsArgs extends BaseFlipArgs {
   unflippedIds: string[]
