+++ packages/flip-toolkit/src/index.ts
@@ -1,7 +1,7 @@
-import * as utilities from './utilities'
-import * as constants from './constants'
-export { default as Flipper } from './Flipper'
-export { default as getFlippedElementPositionsBeforeUpdate } from './flip/getFlippedElementPositions/getFlippedElementPositionsBeforeUpdate'
-export * from './flip'
+import * as utilities from './utilities/index.js'
+import * as constants from './constants.js'
+export { default as Flipper } from './Flipper.js'
+export { default as getFlippedElementPositionsBeforeUpdate } from './flip/getFlippedElementPositions/getFlippedElementPositionsBeforeUpdate/index.js'
+export * from './flip/index.js'
 export { utilities, constants }
-export { default as spring } from './Spring'
+export { default as spring } from './Spring/index.js'
